# specs/001-create-a-rest/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Test Results Management API
  description: |
    REST API for CRUD operations on end-to-end test execution results.
    Primary integration via GitHub Actions CI/CD pipelines using bearer token authentication.

    ## Authentication
    All endpoints require JWT bearer token authentication. Tokens are obtained through
    GitHub SSO for human users or generated for automated systems (CI/CD pipelines).

    ## Integration Patterns
    - **GitHub Actions**: Use bearer tokens stored in repository/organization secrets
    - **CI/CD Systems**: Automated test result submission with artifact upload
    - **Development**: Local testing with personal access tokens
  version: 0.4.0
  contact:
    name: Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.testresults.dev
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestArtifact:
      type: object
      required: [artifact_type, file_name, file_size, mime_type]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        result_id:
          type: string
          format: uuid
          nullable: true
        suite_id:
          type: string
          format: uuid
          nullable: true
        artifact_type:
          type: string
          enum: [screenshot, video, report, log, trace, other]
          example: "screenshot"
        file_name:
          type: string
          maxLength: 255
          example: "login-test-failure.png"
        file_size:
          type: integer
          minimum: 1
          maximum: 104857600
          example: 245760
        mime_type:
          type: string
          example: "image/png"
        storage_key:
          type: string
          readOnly: true
          example: "2025/09/14/001-suite-uuid/screenshot-001.png"
        storage_url:
          type: string
          format: uri
          nullable: true
          readOnly: true
          example: "https://s3.amazonaws.com/test-artifacts/..."
        checksum:
          type: string
          readOnly: true
          example: "sha256:abcd1234..."
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
...
paths:
...
  /auth/me:
    get:
      summary: Get current user info
      tags: [Authentication]
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                    enum: [admin, user, readonly]
                  token_type:
                    type: string
                    enum: [user, automation, ci]
                    description: Type of token being used
        '401':
          description: Authentication required
  /auth/token:
    post:
      summary: Generate automation token
      description: Generate bearer token for CI/CD and automation use
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scope]
              properties:
                name:
                  type: string
                  example: "GitHub Actions - my-repo"
                  description: Descriptive name for the token
                scope:
                  type: string
                  enum: [read, write, admin]
                  description: Token permission scope
                expires_days:
                  type: integer
                  minimum: 1
                  maximum: 365
                  default: 90
                  description: Token expiration in days
                metadata:
                  type: object
                  additionalProperties: true
                  example:
                    repository: "owner/repo"
                    workflow: "test.yml"
      responses:
        '201':
          description: Token created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_id:
                    type: string
                    format: uuid
                  expires_at:
                    type: string
                    format: date-time
                  scope:
                    type: string
                    enum: [read, write, admin]
        '403':
          description: Insufficient permissions to create tokens
...
  /artifacts:
    get:
      summary: List test artifacts
      tags: [Artifacts]
      parameters:
        - name: result_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by test result
        - name: suite_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by test suite
        - name: artifact_type
          in: query
          schema:
            type: string
            enum: [screenshot, video, report, log, trace, other]
          description: Filter by artifact type
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of test artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestArtifact'

    post:
      summary: Upload test artifact
      tags: [Artifacts]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, artifact_type]
              properties:
                file:
                  type: string
                  format: binary
                  description: Artifact file to upload
                result_id:
                  type: string
                  format: uuid
                  description: Associated test result (optional)
                suite_id:
                  type: string
                  format: uuid
                  description: Associated test suite (optional)
                artifact_type:
                  type: string
                  enum: [screenshot, video, report, log, trace, other]
                metadata:
                  type: string
                  description: JSON metadata for the artifact
      responses:
        '201':
          description: Artifact uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestArtifact'
        '400':
          description: Invalid upload parameters
        '413':
          description: File too large (max 100MB)
  /artifacts/{artifact_id}:
    get:
      summary: Get artifact metadata
      tags: [Artifacts]
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Artifact metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestArtifact'
        '404':
          description: Artifact not found

    delete:
      summary: Delete test artifact
      tags: [Artifacts]
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Artifact deleted successfully
        '404':
          description: Artifact not found
...