# src/api/artifacts.py
"""
Artifact API endpoints for test artifact management.
Provides REST API for artifact metadata management (file operations simplified for Phase 6).
"""

from typing import List, Optional
from uuid import UUID

import structlog
from fastapi import APIRouter, Depends, HTTPException, Query, status
from fastapi.responses import Response

from ..lib.middleware import RequireArtifactsRead, RequireArtifactsWrite, RequireArtifactsDelete
from .models import ArtifactResponse

logger = structlog.get_logger()

router = APIRouter(prefix="/api/v1/artifacts", tags=["artifacts"])


@router.get(
    "",
    response_model=List[ArtifactResponse],
    summary="Get artifacts",
    description="Retrieve test artifacts (placeholder for Phase 6)",
    dependencies=[RequireArtifactsRead]
)
async def get_artifacts() -> List[ArtifactResponse]:
    """Get test artifacts (placeholder implementation)."""
    try:
        # Placeholder implementation - return empty list
        logger.debug("Artifacts retrieved via API", count=0)
        return []

    except Exception as e:
        logger.error("Artifact retrieval failed", error=str(e))
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Internal server error occurred while retrieving artifacts"
        )


@router.get(
    "/{artifact_id}",
    response_model=ArtifactResponse,
    summary="Get an artifact by ID",
    description="Retrieve a specific test artifact by its UUID (placeholder for Phase 6)",
    dependencies=[RequireArtifactsRead]
)
async def get_artifact(artifact_id: UUID) -> ArtifactResponse:
    """Get a test artifact by ID (placeholder implementation)."""
    try:
        # Placeholder implementation - always return not found
        logger.warning("Artifact not found via API", artifact_id=str(artifact_id))
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Artifact with ID '{artifact_id}' not found"
        )

    except HTTPException:
        raise
    except Exception as e:
        logger.error("Artifact retrieval failed", artifact_id=str(artifact_id), error=str(e))
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Internal server error occurred while retrieving artifact"
        )